Improved enemy
They dont get stuck anymore on the wall and other object. but the sfm enemy is so close to you, should i put some distance cause his only attack is just a bullet. is like the enemy running to his death and when playing a long period of time he is very annoying. i think that is good though. not sure.