Lockable inventory item
You can only drop item from the tab button(inventory screen) en bij pressing alt and kliking in the item you wanna lock you wil see a 🔒 logo on the item and then that item is not dropable. by pressing alt again and on the same item the lock should be gone, and then you can drop the item.
when the item is locked or not locked you can still pick up the same item and the amount wil still increase.